+++ CHANGELOG.md
@@ -24,3 +24,4 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - fix/1200099491334042 - Fix remove on RemovableInterfaceTrait - Add remove example - Add ArrayHelper
+- fix/1200161406794549 - Fix KeyFilterInterface (and trait)